agent/requirements.txt
langgraph>=0.4.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
openai>=1.0.0
supabase>=2.0.0
pypdf>=4.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
